Patrón de arquitectura MVVM completa.

+++ app/src/main/java/com/ve/tecno/themoviedb/response/Movie.kt
@@ -13,7 +13,7 @@ data class Movie(
     val release_date: String,
     val title: String,
     val video: Boolean,
-    val vote_average: Double,
+    val vote_average: Float,
     val vote_count: Int
 )
